cron job to delete obsolete cache
