+++ chart/templates/_common/_helpers.tpl
@@ -134,6 +134,11 @@ app.kubernetes.io/component: "{{ include "name" . }}-clean-hf-datasets-cache"
 app.kubernetes.io/component: "{{ include "name" . }}-clean-stats-cache"
 {{- end -}}
 
+{{- define "labels.deleteObsoleteCache" -}}
+{{ include "hf.labels.commons" . }}
+app.kubernetes.io/component: "{{ include "name" . }}-delete-obsolete-cache"
+{{- end -}}
+
 {{- define "labels.postMessages" -}}
 {{ include "hf.labels.commons" . }}
 app.kubernetes.io/component: "{{ include "name" . }}-post-messages"
