+++ jobs/cache_maintenance/src/cache_maintenance/main.py
@@ -6,9 +6,11 @@ import sys
 from datetime import datetime
 
 from libcommon.log import init_logging
+from libcommon.obsolete_cache import delete_obsolete_cache
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.storage import init_dir
+from libcommon.storage_client import StorageClient
 
 from cache_maintenance.backfill import backfill_cache
 from cache_maintenance.cache_metrics import collect_cache_metrics
@@ -88,6 +90,30 @@ def run_job() -> None:
                 bot_associated_user_name=job_config.discussions.bot_associated_user_name,
                 bot_token=job_config.discussions.bot_token,
                 parquet_revision=job_config.discussions.parquet_revision,
+            )
+        elif action == "delete-obsolete-cache":
+            cached_assets_storage_client = StorageClient(
+                protocol=job_config.cached_assets.storage_protocol,
+                root=job_config.cached_assets.storage_root,
+                folder=job_config.cached_assets.folder_name,
+                key=job_config.s3.access_key_id,
+                secret=job_config.s3.secret_access_key,
+                client_kwargs={"region_name": job_config.s3.region_name},
+            )
+
+            assets_storage_client = StorageClient(
+                protocol=job_config.assets.storage_protocol,
+                root=job_config.assets.storage_root,
+                folder=job_config.assets.folder_name,
+                key=job_config.s3.access_key_id,
+                secret=job_config.s3.secret_access_key,
+                client_kwargs={"region_name": job_config.s3.region_name},
+            )
+            delete_obsolete_cache(
+                hf_endpoint=job_config.common.hf_endpoint,
+                hf_token=job_config.common.hf_token,
+                cached_assets_storage_client=cached_assets_storage_client,
+                assets_storage_client=assets_storage_client,
             )
         elif action == "skip":
             pass
